+++ R/wflow_open.R
@@ -110,9 +110,9 @@ wflow_open <- function(files,
 
     # Send a warning if user has a beta workflowr project and tries to create a
     # new file
-    yml <- yaml::read_yaml(file.path(p$analysis, "_site.yml"))
-    if (names(yml$output) != "workflowr::wflow_html" &&
-        length(files_new) > 0) {
+    yml_index <- rmarkdown::yaml_front_matter(file.path(p$analysis, "index.Rmd"))
+    usingBeta <- is.null(yml_index[["site"]]) || yml_index[["site"]] != "workflowr::wflow_site"
+    if (usingBeta && length(files_new) > 0) {
       warning(call. = FALSE, wrap(
         "It appears that your site was created using a beta release of
         workflowr, thus you likely don't want to use the R Markdown file